Split office component
