+++ .vscode/launch.json
@@ -1,0 +1,29 @@
+{
+	"version": "0.2.0",
+	"configurations": [
+		{
+			"name": "Launch index.html",
+			"type": "chrome",
+			"request": "launch",
+			"file": "${workspaceRoot}/index.html"
+		},
+		{
+			"name": "Launch localhost with sourcemaps",
+			"type": "chrome",
+			"request": "launch",
+			"url": "http://localhost:8000/public/index.html",
+            "diagnosticLogging": true,
+			"sourceMaps": true
+		},
+		{
+			"name": "Attach",
+			"type": "chrome",
+			"request": "attach",
+            "diagnosticLogging": true,
+			"port": 9222,
+            "sourceMaps": true,
+            "sourceRoot": "${workspaceRoot}/src",
+            "webRoot": "${workspaceRoot}/public"
+		}
+	]
+}
