+++ src/lib/utilities.ts
@@ -28,8 +28,6 @@ export class Storage {
     }
 }
 
-
-
 export function Hook(rootObject: any, functionToHook: string, hookingFunction: (...optionalParams: any[]) => void): void {
     var previousFunction = rootObject[functionToHook];
 
@@ -37,9 +35,9 @@ export function Hook(rootObject: any, functionToHook: string, hookingFunction: (
         hookingFunction.apply(null,optionalParams);
         previousFunction.apply(rootObject, optionalParams);
     }
+
     return previousFunction;
 }
-
 
 export function CreateCookie(name: string, value: string, days: number): void {
     var expires: string;
@@ -113,22 +111,20 @@ export function sortBy(key?: (any) => any, reverse?: boolean) {
 }
 
 export class DebugConsole {
-    
     console : HTMLDivElement;
     logger : HTMLDivElement;
     command : HTMLInputElement;
-    
 
     public constructor() {
         this.console = document.getElementById("DebugConsole") as HTMLDivElement;
-        this.command = document.getElementById("DebugCommand") as HTMLInputElement; 
+        this.command = document.getElementById("DebugCommand") as HTMLInputElement;
         this.logger = document.getElementById("DebugLog") as HTMLDivElement;
-        
+
         if (!this.console || !this.command || !this.logger) { alert("Unable to initialize local console"); return; }
-        
+
         Hook(console, "log", this.Log );
-        
-        this.console.style.display = "";        
+
+        this.console.style.display = "";
         this.command.onchange = () => {
             var result = ""
             try { result = eval(this.command.value); } catch (err) { result = "Unable to evaluate " + this.command.value + " with error " + err.toString(); }
@@ -137,7 +133,7 @@ export class DebugConsole {
         }
     }
 
-    public Log = (...args : any[]) => {        
+    public Log = (...args : any[]) => {
         var message = args && args[0];
         if (message) {
             try {
@@ -153,15 +149,13 @@ export class DebugConsole {
 
 export var LocalConsole = null;
 
-export function LocalConsoleInitialize()
-{
+export function LocalConsoleInitialize() {
     LocalConsole = new DebugConsole();
 }
 
 export enum Days { Mon = 1, Tue, Wed, Thu, Fri, Sat, Sun }
 
-export function ShortTimeString(dateString: string)
-{
+export function ShortTimeString(dateString: string) {
     var today = new Date();
     var date = new Date(dateString);
 
@@ -172,7 +166,7 @@ export function ShortTimeString(dateString: string)
             hours = 12;
         } else if (hours >= 12) {
             suffix = " PM";
-            hours -= 12;       
+            hours -= 12;
         }
         var minutes = date.getMinutes().toString();
         if (minutes.length == 1) {
@@ -180,6 +174,6 @@ export function ShortTimeString(dateString: string)
         }
         return hours + ":" + minutes + suffix;
     } else {
-        return Days[date.getDay()] + " " + date.getMonth() + "/" + date.getDate();        
+        return Days[date.getDay()] + " " + date.getMonth() + "/" + date.getDate();
     }
 }
